fix: Refactor createRecord test for Rest adapter
