+++ packages/-ember-data/tests/integration/adapter/rest-adapter-test.js
@@ -131,380 +131,6 @@ module('integration/adapter/rest_adapter - REST Adapter', function (hooks) {
     };
   }
 
-  test('createRecord - an empty payload is a basic success if an id was specified', async function (assert) {
-    ajaxResponse();
-
-    let post = store.createRecord('post', { id: 'some-uuid', name: 'The Parley Letter' });
-    await post.save();
-    assert.equal(passedUrl, '/posts');
-    assert.equal(passedVerb, 'POST');
-    assert.deepEqual(passedHash.data, { post: { id: 'some-uuid', name: 'The Parley Letter' } });
-
-    assert.false(post.get('hasDirtyAttributes'), "the post isn't dirty anymore");
-    assert.equal(post.get('name'), 'The Parley Letter', 'the post was updated');
-  });
-
-  test('createRecord - passes buildURL the requestType', async function (assert) {
-    adapter.buildURL = function (type, id, snapshot, requestType) {
-      return '/post/' + requestType;
-    };
-
-    ajaxResponse();
-
-    let post = store.createRecord('post', { id: 'some-uuid', name: 'The Parley Letter' });
-    await post.save();
-    assert.equal(passedUrl, '/post/createRecord');
-  });
-
-  test('createRecord - a payload with a new ID and data applies the updates', async function (assert) {
-    ajaxResponse({ posts: [{ id: '1', name: 'Dat Parley Letter' }] });
-
-    let post = store.createRecord('post', { name: 'The Parley Letter' });
-
-    await post.save();
-    assert.equal(passedUrl, '/posts');
-    assert.equal(passedVerb, 'POST');
-    assert.deepEqual(passedHash.data, { post: { name: 'The Parley Letter' } });
-
-    assert.equal(post.get('id'), '1', 'the post has the updated ID');
-    assert.false(post.get('hasDirtyAttributes'), "the post isn't dirty anymore");
-    assert.equal(post.get('name'), 'Dat Parley Letter', 'the post was updated');
-  });
-
-  test('createRecord - a payload with a new ID and data applies the updates (with legacy singular name)', async function (assert) {
-    ajaxResponse({ post: { id: '1', name: 'Dat Parley Letter' } });
-    let post = store.createRecord('post', { name: 'The Parley Letter' });
-
-    await post.save();
-    assert.equal(passedUrl, '/posts');
-    assert.equal(passedVerb, 'POST');
-    assert.deepEqual(passedHash.data, { post: { name: 'The Parley Letter' } });
-
-    assert.equal(post.get('id'), '1', 'the post has the updated ID');
-    assert.false(post.get('hasDirtyAttributes'), "the post isn't dirty anymore");
-    assert.equal(post.get('name'), 'Dat Parley Letter', 'the post was updated');
-  });
-
-  test("createRecord - findMany doesn't overwrite owner", async function (assert) {
-    ajaxResponse({ comment: { id: '1', name: 'Dat Parley Letter', post: 1 } });
-
-    Post.reopen({ comments: DS.hasMany('comment', { async: true }) });
-    Comment.reopen({ post: DS.belongsTo('post', { async: false }) });
-
-    store.push({
-      data: {
-        type: 'post',
-        id: '1',
-        attributes: {
-          name: 'Rails is omakase',
-        },
-        relationships: {
-          comments: {
-            data: [],
-          },
-        },
-      },
-    });
-    let post = store.peekRecord('post', 1);
-
-    let comment = store.createRecord('comment', { name: 'The Parley Letter' });
-
-    post.get('comments').pushObject(comment);
-
-    assert.equal(comment.get('post'), post, 'the post has been set correctly');
-
-    await comment.save();
-    assert.false(comment.get('hasDirtyAttributes'), "the post isn't dirty anymore");
-    assert.equal(comment.get('name'), 'Dat Parley Letter', 'the post was updated');
-    assert.equal(comment.get('post'), post, 'the post is still set');
-  });
-
-  test("createRecord - a serializer's primary key and attributes are consulted when building the payload", async function (assert) {
-    this.owner.register(
-      'serializer:post',
-      DS.RESTSerializer.extend({
-        primaryKey: '_id_',
-
-        attrs: {
-          name: '_name_',
-        },
-      })
-    );
-
-    ajaxResponse();
-
-    let post = store.createRecord('post', { id: 'some-uuid', name: 'The Parley Letter' });
-
-    await post.save();
-    assert.deepEqual(passedHash.data, {
-      post: { _id_: 'some-uuid', _name_: 'The Parley Letter' },
-    });
-  });
-
-  test("createRecord - a serializer's attributes are consulted when building the payload if no id is pre-defined", async function (assert) {
-    let post;
-    this.owner.register(
-      'serializer:post',
-      DS.RESTSerializer.extend({
-        attrs: {
-          name: '_name_',
-        },
-      })
-    );
-
-    ajaxResponse({
-      post: { _name_: 'The Parley Letter', id: '1' },
-    });
-
-    post = store.createRecord('post', { name: 'The Parley Letter' });
-
-    await post.save();
-    assert.deepEqual(passedHash.data, { post: { _name_: 'The Parley Letter' } });
-  });
-
-  test("createRecord - a serializer's attribute mapping takes precedence over keyForAttribute when building the payload", async function (assert) {
-    this.owner.register(
-      'serializer:post',
-      DS.RESTSerializer.extend({
-        attrs: {
-          name: 'given_name',
-        },
-
-        keyForAttribute(attr) {
-          return attr.toUpperCase();
-        },
-      })
-    );
-
-    ajaxResponse();
-
-    let post = store.createRecord('post', { id: 'some-uuid', name: 'The Parley Letter' });
-
-    await post.save();
-    assert.deepEqual(passedHash.data, {
-      post: { given_name: 'The Parley Letter', id: 'some-uuid' },
-    });
-  });
-
-  test("createRecord - a serializer's attribute mapping takes precedence over keyForRelationship (belongsTo) when building the payload", async function (assert) {
-    this.owner.register(
-      'serializer:comment',
-      DS.RESTSerializer.extend({
-        attrs: {
-          post: 'article',
-        },
-
-        keyForRelationship(attr, kind) {
-          return attr.toUpperCase();
-        },
-      })
-    );
-
-    ajaxResponse();
-
-    Comment.reopen({ post: DS.belongsTo('post', { async: false }) });
-
-    let post = store.createRecord('post', { id: 'a-post-id', name: 'The Parley Letter' });
-    let comment = store.createRecord('comment', {
-      id: 'some-uuid',
-      name: 'Letters are fun',
-      post: post,
-    });
-
-    await comment.save();
-    assert.deepEqual(passedHash.data, {
-      comment: { article: 'a-post-id', id: 'some-uuid', name: 'Letters are fun' },
-    });
-  });
-
-  test("createRecord - a serializer's attribute mapping takes precedence over keyForRelationship (hasMany) when building the payload", async function (assert) {
-    this.owner.register(
-      'serializer:post',
-      DS.RESTSerializer.extend({
-        attrs: {
-          comments: 'opinions',
-        },
-
-        keyForRelationship(attr, kind) {
-          return attr.toUpperCase();
-        },
-      })
-    );
-
-    ajaxResponse();
-
-    Post.reopen({ comments: DS.hasMany('comment', { async: false }) });
-
-    let comment = store.createRecord('comment', { id: 'a-comment-id', name: 'First!' });
-    let post = store.createRecord('post', {
-      id: 'some-uuid',
-      name: 'The Parley Letter',
-      comments: [comment],
-    });
-
-    await post.save();
-    assert.deepEqual(passedHash.data, {
-      post: { opinions: ['a-comment-id'], id: 'some-uuid', name: 'The Parley Letter' },
-    });
-  });
-
-  test('createRecord - a record on the many side of a hasMany relationship should update relationships when data is sideloaded', async function (assert) {
-    assert.expect(3);
-
-    ajaxResponse({
-      posts: [
-        {
-          id: '1',
-          name: 'Rails is omakase',
-          comments: [1, 2],
-        },
-      ],
-      comments: [
-        {
-          id: '2',
-          name: 'Another Comment',
-          post: 1,
-        },
-        {
-          id: '1',
-          name: 'Dat Parley Letter',
-          post: 1,
-        },
-      ],
-      // My API is returning a comment:{} as well as a comments:[{...},...]
-      //, comment: {
-      //   id: "2",
-      //   name: "Another Comment",
-      //   post: 1
-      // }
-    });
-
-    Post.reopen({ comments: DS.hasMany('comment', { async: false }) });
-    Comment.reopen({ post: DS.belongsTo('post', { async: false }) });
-
-    store.push({
-      data: {
-        type: 'post',
-        id: '1',
-        attributes: {
-          name: 'Rails is omakase',
-        },
-        relationships: {
-          comments: {
-            data: [{ type: 'comment', id: '1' }],
-          },
-        },
-      },
-    });
-    store.push({
-      data: {
-        type: 'comment',
-        id: '1',
-        attributes: {
-          name: 'Dat Parlay Letter',
-        },
-        relationships: {
-          post: {
-            data: { type: 'post', id: '1' },
-          },
-        },
-      },
-    });
-
-    let post = store.peekRecord('post', 1);
-    let commentCount = post.get('comments.length');
-
-    assert.equal(commentCount, 1, 'the post starts life with a comment');
-
-    let comment = store.createRecord('comment', { name: 'Another Comment', post: post });
-    await comment.save();
-    assert.equal(comment.get('post'), post, 'the comment is related to the post');
-    await post.reload();
-    assert.equal(post.get('comments.length'), 2, 'Post comment count has been updated');
-  });
-
-  test('createRecord - sideloaded belongsTo relationships are both marked as loaded', async function (assert) {
-    assert.expect(4);
-
-    Post.reopen({ comment: DS.belongsTo('comment', { async: false }) });
-    Comment.reopen({ post: DS.belongsTo('post', { async: false }) });
-
-    let post = store.createRecord('post', { name: 'man' });
-
-    ajaxResponse({
-      posts: [{ id: 1, comment: 1, name: 'marked' }],
-      comments: [{ id: 1, post: 1, name: 'Comcast is a bargain' }],
-    });
-
-    let record = await post.save();
-    assert.true(store.peekRecord('post', 1).get('comment.isLoaded'), "post's comment isLoaded (via store)");
-    assert.true(store.peekRecord('comment', 1).get('post.isLoaded'), "comment's post isLoaded (via store)");
-    assert.true(record.get('comment.isLoaded'), "post's comment isLoaded (via record)");
-    assert.true(record.get('comment.post.isLoaded'), "post's comment's post isLoaded (via record)");
-  });
-
-  test("createRecord - response can contain relationships the client doesn't yet know about", async function (assert) {
-    assert.expect(3); // while records.length is 2, we are getting 4 assertions
-
-    ajaxResponse({
-      posts: [
-        {
-          id: '1',
-          name: 'Rails is omakase',
-          comments: [2],
-        },
-      ],
-      comments: [
-        {
-          id: '2',
-          name: 'Another Comment',
-          post: 1,
-        },
-      ],
-    });
-
-    Post.reopen({ comments: DS.hasMany('comment', { async: false }) });
-    Comment.reopen({ post: DS.belongsTo('post', { async: false }) });
-
-    let post = store.createRecord('post', { name: 'Rails is omakase' });
-
-    await post.save();
-    assert.equal(post.get('comments.firstObject.post'), post, 'the comments are related to the correct post model');
-    assert.equal(
-      store._internalModelsFor('post').models.length,
-      1,
-      'There should only be one post record in the store'
-    );
-
-    let postRecords = store._internalModelsFor('post').models;
-    for (var i = 0; i < postRecords.length; i++) {
-      assert.equal(post, postRecords[i].getRecord(), 'The object in the identity map is the same');
-    }
-  });
-
-  test('createRecord - relationships are not duplicated', async function (assert) {
-    Post.reopen({ comments: DS.hasMany('comment', { async: false }) });
-    Comment.reopen({ post: DS.belongsTo('post', { async: false }) });
-
-    let post = store.createRecord('post', { name: 'Tomtomhuda' });
-    let comment = store.createRecord('comment', { id: 2, name: 'Comment title' });
-
-    ajaxResponse({ post: [{ id: 1, name: 'Rails is omakase', comments: [] }] });
-
-    await post.save();
-    assert.equal(post.get('comments.length'), 0, 'post has 0 comments');
-    post.get('comments').pushObject(comment);
-    assert.equal(post.get('comments.length'), 1, 'post has 1 comment');
-
-    ajaxResponse({
-      post: [{ id: 1, name: 'Rails is omakase', comments: [2] }],
-      comments: [{ id: 2, name: 'Comment title' }],
-    });
-
-    await post.save();
-    assert.equal(post.get('comments.length'), 1, 'post has 1 comment');
-  });
-
   test('updateRecord - an empty payload is a basic success', async function (assert) {
     store.push({
       data: {
